GoogleSearchUriInput: add instrumented test
