+++ app/src/androidTest/java/page/ooooo/geoshare/tests/inputs/CartesIGNInputBehaviorTest.kt
@@ -6,7 +6,7 @@ import page.ooooo.geoshare.lib.geo.Source
 import page.ooooo.geoshare.lib.geo.WGS84Point
 import page.ooooo.geoshare.tests.testUri
 
-class CartesIGNInputBehaviorTest {
+class CartesIGNInputBehaviorTest : InputBehaviorTest {
     @Test
     fun cartesIGN_offline() = uiAutomator {
         // Coordinates
