+++ app/src/androidTest/java/page/ooooo/geoshare/tests/inputs/HereWeGoInputBehaviorTest.kt
@@ -6,7 +6,7 @@ import page.ooooo.geoshare.lib.geo.Source
 import page.ooooo.geoshare.lib.geo.WGS84Point
 import page.ooooo.geoshare.tests.testUri
 
-class HereWeGoInputBehaviorTest {
+class HereWeGoInputBehaviorTest : InputBehaviorTest {
     @Test
     fun hereWeGo_offline() = uiAutomator {
         // Coordinates
