+++ app/src/androidTest/java/page/ooooo/geoshare/tests/inputs/MagicEarthInputBehaviorTest.kt
@@ -6,7 +6,7 @@ import page.ooooo.geoshare.lib.geo.Source
 import page.ooooo.geoshare.lib.geo.WGS84Point
 import page.ooooo.geoshare.tests.testUri
 
-class MagicEarthInputBehaviorTest {
+class MagicEarthInputBehaviorTest : InputBehaviorTest {
     @Test
     fun magicEarth_offline() = uiAutomator {
         // Coordinates
